finalizing changes, added 3 additional models, don't have time to display the information on the client side, on the server side, the resulting outputs and graphs are stored in the Text_Files and visualization directories, will need to work on pitch and submission now
