+++ server/Machine-Learning/Text_Files/Antalya_Dental_Clinics_VADER_sentiment.txt
@@ -1,0 +1,8 @@
+Skipping a review that is not in the expected format.
+
+Skipping a review that is not in the expected format.
+
+Skipping a review that is not in the expected format.
+
+Skipping a review that is not in the expected format.
+
