+++ server/Machine-Learning/Text_Files/Attelia_Oral_and_Dental_Health_Center_VADER_sentiment.txt
@@ -1,0 +1,70 @@
+Review: My experience of Attelia has been nothing short of AMAZING!!
+
+From the first pick-up at the airport back in January, everything has been run like clockwork and so precisely on time.
+
+My first visit in January was to get the implants inserted and this was all done perfectly - the aftercare was also just superb!
+
+The second visit to have the crowns fitted has also gone just as the first visit did; All gone to clockwork and on time.
+
+The teeth I have had inserted look and feel great, I couldn't have asked for a bettwr experience or service.
+
+My coordinator Kagan and specialist Ozer on the second visit have been absolutely great; They have made sure I have felt comfortable and informed throughout the whole process.
+
+Conclusion from myself:
+
+If you want to have work done on your teeth and are apprehensive or have any worries, choose ATTELIA and they will meet all your expectations and more!!     ❤️
+Polarity Scores: {'neg': 0.035, 'neu': 0.816, 'pos': 0.149, 'compound': 0.968}
+
+Review: Dear Attelia, thank you so much for another amazing experience at the clinic. All the staff here make the experience so much nicer and for sure the new Bar area is a great addition, especially if you have to spend all day here.
+It truly is the best dental clinic in the world and I have recommended many to come to Attelia to get their perfect smile. All have said how amazing everyone is and the overall experience is so great.
+Keep up the amazing work and see you in a few months for my implants.
+Take care until then.
+Polarity Scores: {'neg': 0.0, 'neu': 0.631, 'pos': 0.369, 'compound': 0.9938}
+
+Review: Well …. Not sure where to start. First of all I don’t like travelling or leaving home very much never mind skipping off to another country so far away for dental work.
+I must also say that I am working class, I’m not used to investing money in my looks.
+So how do you pick a clinic. My choice and I believe most people’s choice was purely based on knowing someone that has been here before.
+I mean it, and I’m also not much of a “reviewer” …… this place, this service, they staff ….. everything just feels comfortable.
+The professionalism, the cleanliness, the friendliness ……. It just all flows!
+I’ve never enjoyed the dentists and sure who does. I thought this would be my last horrendous dentist visit …….. but for the first time in my life I felt comfortable here.
+It runs so smoothly ……. Like a boat through water …….. and I like boats ….. a lot!
+Polarity Scores: {'neg': 0.033, 'neu': 0.848, 'pos': 0.12, 'compound': 0.936}
+
+Review: From the initial contact with Attelia, their professionalism has been consistently impressive. I was promptly provided with a comprehensive plan detailing the procedures, ensuring transparency and clarity. The experience continued to exceed expectations as I was chauffeured in a luxurious car upon my arrival at the airport, with punctual pick-ups every day.
+
+The meticulous organization extended to providing my appointment times a day in advance, leaving no room for dissatisfaction. The dental expertise at Attelia is truly world-class, creating an environment where one feels secure and heard, addressing any concerns with utmost care.
+
+Throughout my stay, my interpreter played a pivotal role in ensuring I remained well-informed and content with every aspect of the process. I am now elated to have achieved the smile I had long desired, and I cannot express enough gratitude for Attelia's role in making it a reality.
+
+I wholeheartedly recommend Attelia to anyone seeking dental work at an affordable price without compromising on a top-notch, professional service. Thank you, Attelia, for transforming the smile I dreamt of into a reality.
+Polarity Scores: {'neg': 0.039, 'neu': 0.758, 'pos': 0.202, 'compound': 0.9826}
+
+Review: Great service, very high tech and modern clinic. Burak and Selvi was great at communicating each stage of the treatment and in regards to travelling. I have had four fillings done and descale and polish. The snack and drinks are all for free too. Place looks amazing! You get a copy of the treatment in a Attelia bag at the end. ONLY issue is you may have to wait for the taxi for a bit as you have to bug the information desk about the taxi timing. I just hope the fillings stay all intact.
+Polarity Scores: {'neg': 0.0, 'neu': 0.814, 'pos': 0.186, 'compound': 0.9652}
+
+Review: Amazing dental clinic in turkey, can’t speak highly enough about it. Best experience I’ve ever had at a dental clinic, everybody is so friendly and professional. Exceptional standards. I’d also like to give a special mention to Can and his assistant. Can is a fantastic dentist with unreal attention to detail and a perfectionist who wanted the finished job to be of the highest standard. Thank you.
+Polarity Scores: {'neg': 0.0, 'neu': 0.695, 'pos': 0.305, 'compound': 0.9774}
+
+Review: Amazing. From you walk in the door everyone is friendly and attentive.  A very professional clinic.  I had 4 extractions for my first visit.  It was virtually painless.  The attention to detail and explanation of the treatment this and future visit was thorough.  I can't recommend enough. Would definitely recommend.  I will be attending this clinic for all future dentistry. The premises is modern and clean.
+Polarity Scores: {'neg': 0.027, 'neu': 0.754, 'pos': 0.219, 'compound': 0.9324}
+
+Review: Dear Attila Clinic Team,
+
+I wanted to take a moment to express my sincere gratitude to you all for the exceptional dental care that I recently received at your clinic.
+From the moment I walked through your doors, I was met with nothing but kindness and professionalism. Your staff made me feel comfortable and well-informed about every aspect of my treatment. The attention to detail, patience, and genuine care exhibited by each team member did not go unnoticed and was very much appreciated.
+I would like to extend a special thank you to the doctor who treated me. Their expertise and skill have truly made a difference in my life. My teeth have never felt better, and I am so thankful for the work they did.
+Implantologist Mr Hamit and my translator and coordinator Selvi
+I also appreciated the clinic's clean and welcoming environment, which made my visits much more comfortable than I could have anticipated.
+Your dedication to your patients' health and well-being is apparent in every interaction, and I feel fortunate to have been under your care. Thank you once again for the incredible work you all do.
+
+Best Regards,
+
+Pop Andrei Ioan
+Polarity Scores: {'neg': 0.025, 'neu': 0.671, 'pos': 0.305, 'compound': 0.9973}
+
+Review: By far the best dentist I have ever visited with the best equipment and facilities. Same excellent treatment and hospitality as usual as we completed my treatment, as planned, 6 months after it began.  My coordinator Kagan and dentist Ahmet listened carefully to all my questions and took me through the process step by step. Highly recommended.
+Polarity Scores: {'neg': 0.0, 'neu': 0.768, 'pos': 0.232, 'compound': 0.9402}
+
+Review: Very happy with my veneers. I found Attelia on Youtube. The treatment itself went well. There was no language issue; all staff members were able to understand and accommodate my dental requests. Even when I needed smaller tweaks to my veneers the dentist, assistants, and designer/artist was patient. The treatment recommended to me was also well explained. The hotel arrangement and transportation was well organized. The office itself was also very modern and clean. I had the privilege to be the patient of Dr. Derya Parpar & I got to meet the very knowledgeable Professor Siddik from University of Antalya. I got the natural look, with rounded edges to my teeth to make them look more youthful.
+Polarity Scores: {'neg': 0.017, 'neu': 0.812, 'pos': 0.172, 'compound': 0.9549}
+
