+++ server/Machine-Learning/Text_Files/Cosmetic_Dental_Turkey_sentiment.txt
@@ -1,0 +1,3 @@
+Review: Excellent dentist , friendly , efficient and extremely Professional .  The team put you at easy and offer a VIP experience. I would recommend to everyone!
+Sentiment: Sentiment(polarity=0.5041666666666667, subjectivity=0.6083333333333334)
+
