+++ server/Machine-Learning/Text_Files/Dental_Clinic_Antalya_sentiment.txt
@@ -1,0 +1,45 @@
+Review: Amazing clinic with very kind, caring and professional staff. Lady called Ilke was an  absolut angel throughout the process. She gave my husband consultation online, explained everything, guided him while in Turkey, was super helpful, understanding and supportive since he has needle fear and is dyslexic. She made us feel comfortable and at ease.
+Biggest thank You to all Your team one more time. Will be recommending Your clinic to all our  friends and clients.
+
+Kind regards,
+M and L Morrison
+Sentiment: Sentiment(polarity=0.4766666666666667, subjectivity=0.7333333333333334)
+
+Review: The teeth whitening process put my husband in a lot of pain after it, the results of the 2 sessions combined at 40 mins were not as effective as advertised in their photo. Almost looks like unreal photos used on their pages, there was also some bleeding from his gums, the first photo is the before photo on the same day, the 2 after are straight after the teeth whitening process. Better teeth whitening in London and cheaper, avoid if you can.
+Sentiment: Sentiment(polarity=0.31, subjectivity=0.43166666666666664)
+
+Review: I am extremely pleased with the results I achieved! The staff at the clinic are incredibly friendly and welcoming, and the clinic itself is impeccably clean. I am definitely planning on returning for more treatments in the future xx
+Sentiment: Sentiment(polarity=0.3111111111111111, subjectivity=0.5541666666666667)
+
+Review: My experience from the start was good. The dentists and management were extremely excellent. I and my wife had a total of 12 fillings and 4 crowns and done perfect. The interpreter named ilke looked after us from start to finish very polite and helpful. I would definitely recommend this clinic.
+Sentiment: Sentiment(polarity=0.4833333333333334, subjectivity=0.6916666666666668)
+
+Review: I have just gotten 24 crowns done, and i couldn’t be more happy. This clinic made me feel comfortable and safe about the decision, as they are most certainly experts in the industry, and are able to answer any question or concern. Coming from Canada, people will tell you to steer away from visiting clinics in foreign countries like this, but i can assure you that the dentist’s at this clinic are just as good, if not better then any dentist i have been to in Canada. With their package deal they provided accommodation and transportation from airport, hotel, and clinic, which was a huge plus. If you are looking to get crowns or veneers done i would highly recommend coming here as you won’t find a better service for the price. Everyone was super friendly from the translator, dentist’s, assistants, and receptionist. Thank you again Dental Clinic Turkey!
+Sentiment: Sentiment(polarity=0.37342460317460324, subjectivity=0.5885396825396826)
+
+Review: Professional and the best Dental Clinic.
+I have witnessed and enjoyed services of dentists with highest level of expertise.
+100% hygiene
+100% perfection
+100% customer service and comfortable service.
+100% customer satisfaction
+The supporting staff are with great understanding , caring and responsible.
+Wonderful experience,  value for money.... Exceptional services.
+Happy smiles       
+
+I recommend.
+Sentiment: Sentiment(polarity=0.5716666666666667, subjectivity=0.645)
+
+Review: I had a great experience with this clinic. From the beginning to the end the clinic provides  a warm hospitality and care. The clinic also provides a hotel and a transfer from airport to the hotel , to the clinic and back to the airport.The accommodation is in a walking distance from the clinic, which is very helpful in case if you need to discuss anything with clinic staff. The dentist and surgeon are very professional, quick and attentive. I have 28 all zirconium crowns  and 4 implants been done in 4 days at this clinic and I am very happy with the result! Will be back in a five months to finish the dental work due to implants healing time. I highly recommend this clinic to everyone who’s looking for  a trustworthy dentists who can take care of your teeth and a beautiful “million dollar” smile! Marina Rhoads
+Sentiment: Sentiment(polarity=0.36333333333333334, subjectivity=0.47653846153846147)
+
+Review: I would like to recommend this highly professional clinic in Antalya, I was worried at first , cause there is so many to choose from , but after getting in touch with clinics  coardination team , my worries disappeared. They are very professional and reassured that any needs I might have they will be able to help me and created treatment plan tailored to my individual needs . Its fairly  small clinic and everyone is treated and looked after very well to compare with similar but much bigger clinics. In Dental Clinic Antalya  they are after quality not quantity, they make sure that you leave clinic satisfied and happy . My dentist was truly master of his craft and offered different treatmnents and options to suit my needs and financial possibilities. Will definitely come back soon .Thank you
+Sentiment: Sentiment(polarity=0.18999999999999997, subjectivity=0.45160130718954244)
+
+Review: I came over from Philadelphia and the process was extremely smooth from the online consultation to completion! The staff was very friendly and the dentist had extreme attention to detail. You already know how cost effective it was compared to the US and the results are life changing. Don’t hesitate to get a new smile with them. During the week I explored Antalya and its beauty and left with a perfect smile. You can’t beat that!
+Sentiment: Sentiment(polarity=0.36376262626262623, subjectivity=0.5116161616161616)
+
+Review: finished the first stage of treatment with Dental Clinic Antalya. Came from LONDON .My next visit will be in four or five months later. It was a great experience for me. The experienced and responsible medical team,  most importantly was for me the expertise and honesty of the surgeon، I noticed the surgeon's expertise and skill in implant placement.Because I had the experience of having implants in the past.The clinic is very modern and clean ,The management of receiving treatment services between the patient and the clinic was very precise and without any time wastage. And of course, my translator was compassionate and responsible during the first stage in four days. and HOTEL I had a good time because it was very clean ، next to the beautiful beach of Antalya,t and the staff were polite I really enjoyed being there.
+I personally recommend to my friends and family who need dental treatment GET IT IN DENTAL CLINIC ANTALYA
+Sentiment: Sentiment(polarity=0.3536666666666667, subjectivity=0.5258333333333333)
+
