+++ server/Machine-Learning/Text_Files/Global_Dental_Center_VADER_sentiment.txt
@@ -1,0 +1,54 @@
+Review: Absolutely brilliant!
+I Would 100% recommend the Global Dental Center clinic as they have helped me regain my confidence.
+My smile is just amazing and The team is brilliant from start to finish.
+Top quality service in Turkey.
+You should all Go with Global Dental Center
+Polarity Scores: {'neg': 0.0, 'neu': 0.638, 'pos': 0.362, 'compound': 0.9686}
+
+Review: A lot of thanks to Global Dental Center.
+I am a very happy client here. My case was not easy but Dr Musa and Dr Ayse did a brilliant job. The have dealt with all my issues very professionally and the treatment I have got is high quality. For that, I am very very grateful and I will definitely recommend Global Dental Center to any future customers and anyone. If you want a new smile and new confidence you should come here.
+Polarity Scores: {'neg': 0.016, 'neu': 0.691, 'pos': 0.292, 'compound': 0.983}
+
+Review: I went to Global Dental Center and when Ingot there I was so nervous at first. At a point I even considered going back but once I got to meet Charles and the team, I couldn’t have felt more comfortable. They were very reassuring and they’ve supported me throughout every step of the way. The team was actually amazing with everything they’ve done.
+I couldn’t be happier with my smile and I will reconned anyone who is thinking of going to globe dentist should definitely consider it I can not thank them enuf they went above and beyond for me and made me feel very welcome and safe!
+Polarity Scores: {'neg': 0.033, 'neu': 0.698, 'pos': 0.269, 'compound': 0.989}
+
+Review: I had a fantastic experience.
+I look 10 years younger and I highly recommend Global Dental Center to anyone that wants a good smile.
+The result has increased my confidence and thank you Global Dental Center.
+Polarity Scores: {'neg': 0.0, 'neu': 0.596, 'pos': 0.404, 'compound': 0.9565}
+
+Review: I have visited the Global Dental Centre in Antalya on two occasions to have dental implants.
+Before coming here I was in contact with Charles who was my liason who was outstanding in explaining everything prior to my visit. On arrival at the centre I was informed in depth about my procedures which were performed beyond reproach all the staff from cleaners to dentists were friendly and I cannot thank them enough.
+I would recovery Global Dental Centre to anyone they have reinstalled my confidence.
+Thank you so much
+Polarity Scores: {'neg': 0.021, 'neu': 0.848, 'pos': 0.131, 'compound': 0.8977}
+
+Review: Had my tooth extracted by Dr. Musa Kartoz in Global Dental Center. It was good experience, the doctor informed me about all the process, gave me enough anesthesia to feel no pain. The price was told me before the procedure and there were no additional costs after. Well done.
+Polarity Scores: {'neg': 0.04, 'neu': 0.82, 'pos': 0.14, 'compound': 0.6707}
+
+Review: I am writing to highly recommend Global dental center for any dental services that may be required. As a patient of Global dental center for the past year, I can attest to the high quality of care provided by the team.
+
+From routine checkups to more complex procedures, the dental professionals Global dental center are knowledgeable, skilled, and always put their patients first. They provide a welcoming and comfortable environment for all, making even the most nervous patients feel at ease.
+
+I have always been impressed by the state-of-the-art technology used by the clinic, which ensures efficient and accurate diagnoses and treatments. Additionally, the staff is friendly, efficient, and always go above and beyond to ensure their patients are satisfied with their experience.
+
+Overall, I highly recommend Global dental center for any dental needs. They are truly committed to providing the highest level of care and treating their patients with the utmost respect and professionalism.
+
+Big Thanks to all team !!! :)
+
+Yours Aistijus  (Ice)
+
+London
+Polarity Scores: {'neg': 0.012, 'neu': 0.727, 'pos': 0.261, 'compound': 0.9934}
+
+Review: Great experience with Global Dental Centre.
+Would definitely recommend to anyone travelling from the UK to get their perfect smile : )
+Polarity Scores: {'neg': 0.0, 'neu': 0.523, 'pos': 0.477, 'compound': 0.9382}
+
+Review: I had a great experience with the doctors who treated my jaw fractures. They were always professional, polite, and spoke English fluently. I'm grateful for their help during my six-month recovery, which went smoothly. Thank you, doctors! I hope to see you again soon.
+Polarity Scores: {'neg': 0.0, 'neu': 0.716, 'pos': 0.284, 'compound': 0.9381}
+
+Review: Amazing dental practice. Well spoken English.  The best in Turkey. They explain the whole process leaving you stress free with great prices and services. Couldn’t be more happier.
+Polarity Scores: {'neg': 0.062, 'neu': 0.467, 'pos': 0.471, 'compound': 0.9606}
+
