+++ server/Machine-Learning/Text_Files/Global_Dental_Center_sentiment.txt
@@ -1,0 +1,54 @@
+Review: Absolutely brilliant!
+I Would 100% recommend the Global Dental Center clinic as they have helped me regain my confidence.
+My smile is just amazing and The team is brilliant from start to finish.
+Top quality service in Turkey.
+You should all Go with Global Dental Center
+Sentiment: Sentiment(polarity=0.34444444444444444, subjectivity=0.41111111111111115)
+
+Review: A lot of thanks to Global Dental Center.
+I am a very happy client here. My case was not easy but Dr Musa and Dr Ayse did a brilliant job. The have dealt with all my issues very professionally and the treatment I have got is high quality. For that, I am very very grateful and I will definitely recommend Global Dental Center to any future customers and anyone. If you want a new smile and new confidence you should come here.
+Sentiment: Sentiment(polarity=0.17537878787878788, subjectivity=0.37046401515151506)
+
+Review: I went to Global Dental Center and when Ingot there I was so nervous at first. At a point I even considered going back but once I got to meet Charles and the team, I couldn’t have felt more comfortable. They were very reassuring and they’ve supported me throughout every step of the way. The team was actually amazing with everything they’ve done.
+I couldn’t be happier with my smile and I will reconned anyone who is thinking of going to globe dentist should definitely consider it I can not thank them enuf they went above and beyond for me and made me feel very welcome and safe!
+Sentiment: Sentiment(polarity=0.29038461538461535, subjectivity=0.39487179487179486)
+
+Review: I had a fantastic experience.
+I look 10 years younger and I highly recommend Global Dental Center to anyone that wants a good smile.
+The result has increased my confidence and thank you Global Dental Center.
+Sentiment: Sentiment(polarity=0.156, subjectivity=0.24400000000000005)
+
+Review: I have visited the Global Dental Centre in Antalya on two occasions to have dental implants.
+Before coming here I was in contact with Charles who was my liason who was outstanding in explaining everything prior to my visit. On arrival at the centre I was informed in depth about my procedures which were performed beyond reproach all the staff from cleaners to dentists were friendly and I cannot thank them enough.
+I would recovery Global Dental Centre to anyone they have reinstalled my confidence.
+Thank you so much
+Sentiment: Sentiment(polarity=0.15357142857142855, subjectivity=0.29642857142857143)
+
+Review: Had my tooth extracted by Dr. Musa Kartoz in Global Dental Center. It was good experience, the doctor informed me about all the process, gave me enough anesthesia to feel no pain. The price was told me before the procedure and there were no additional costs after. Well done.
+Sentiment: Sentiment(polarity=0.15, subjectivity=0.30000000000000004)
+
+Review: I am writing to highly recommend Global dental center for any dental services that may be required. As a patient of Global dental center for the past year, I can attest to the high quality of care provided by the team.
+
+From routine checkups to more complex procedures, the dental professionals Global dental center are knowledgeable, skilled, and always put their patients first. They provide a welcoming and comfortable environment for all, making even the most nervous patients feel at ease.
+
+I have always been impressed by the state-of-the-art technology used by the clinic, which ensures efficient and accurate diagnoses and treatments. Additionally, the staff is friendly, efficient, and always go above and beyond to ensure their patients are satisfied with their experience.
+
+Overall, I highly recommend Global dental center for any dental needs. They are truly committed to providing the highest level of care and treating their patients with the utmost respect and professionalism.
+
+Big Thanks to all team !!! :)
+
+Yours Aistijus  (Ice)
+
+London
+Sentiment: Sentiment(polarity=0.1794675925925926, subjectivity=0.3643209876543209)
+
+Review: Great experience with Global Dental Centre.
+Would definitely recommend to anyone travelling from the UK to get their perfect smile : )
+Sentiment: Sentiment(polarity=0.43333333333333335, subjectivity=0.5583333333333333)
+
+Review: I had a great experience with the doctors who treated my jaw fractures. They were always professional, polite, and spoke English fluently. I'm grateful for their help during my six-month recovery, which went smoothly. Thank you, doctors! I hope to see you again soon.
+Sentiment: Sentiment(polarity=0.35, subjectivity=0.3375)
+
+Review: Amazing dental practice. Well spoken English.  The best in Turkey. They explain the whole process leaving you stress free with great prices and services. Couldn’t be more happier.
+Sentiment: Sentiment(polarity=0.5, subjectivity=0.5214285714285715)
+
