+++ server/Machine-Learning/Text_Files/Kotul_Healthcare_VADER_sentiment.txt
@@ -1,0 +1,31 @@
+Review: I love my experience here. The doctor was so nice. There was a sign saying my name when I arrived at the airport. I came from Houston, Texas. They took me to my hotel with their fancy transport. Their coordinator Muhammad was easy to understand and very nice. He told me the entire process and quote. Everything was understood with ease. Yasmin was also very sweet. I'm so glad I chose to come here for a better price than other places. I recommend this place to anyone to get their teeth done. I love my zirconium crowns. I have the smile of my dreams thanks to kotul dental center.
+Polarity Scores: {'neg': 0.0, 'neu': 0.703, 'pos': 0.297, 'compound': 0.9903}
+
+Review: I contacted this dental center while I was on my Antalya trip . I applied to convert 12 of my teeth into zirconium teeth. They responded quickly and made my reservation. All the dentists at the dental center were very professional and helped me with even the smallest care and treatment before doing my teeth. I recommend this professional service to everyone
+Polarity Scores: {'neg': 0.0, 'neu': 0.913, 'pos': 0.087, 'compound': 0.6908}
+
+Review: Absolutely amazing experience. I cannot put into words how much the people at kotuls helped me. I feel a different person now and would recommend that anyone needing help or embarrassed by their smile contact kotuls it will change your life. My only regret is I should of done this years ago . Thank you to everyone at kotuls .
+Polarity Scores: {'neg': 0.073, 'neu': 0.73, 'pos': 0.197, 'compound': 0.8399}
+
+Review: They deserve more than 5 stars. Everything was perfect. Highly recommended
+Polarity Scores: {'neg': 0.0, 'neu': 0.608, 'pos': 0.392, 'compound': 0.6997}
+
+Review: Je recommande vivement la clinique Kotul Dental Center!!!  Ils sont d’un professionnalisme, doux et à l’écoute. J’ai eu droit à un service exceptionnel et un travail bien fait. Leur priorité est la satisfaction du patient.
+L’équipe est sympathique et souriante, on se sent en famille.
+Polarity Scores: {'neg': 0.0, 'neu': 0.921, 'pos': 0.079, 'compound': 0.5826}
+
+Review: Kesinlikle 5 yildizdan cok fazlasini hak ediyorlar.implant yaptirmak icin isvecten geldik havalimanindan bizi guleryuzle karsiladilar  transfer hotel  ilgi  mukemmeldi basta doktor sinan bey  ve saime hanim olmak uzere tum guleryuzlu nazik profosyonel calisanlarina cok tesekkur ediyorum islerini  titizlikle profosonelce  yapiyorlar  dis yaptirmak isteyenlere   siddetle tavsiye ediyorum
+Polarity Scores: {'neg': 0.0, 'neu': 0.939, 'pos': 0.061, 'compound': 0.4404}
+
+Review: Bravo à l’équipe de Kotul Pour leur amabilité et leur professionnalisme, nous avons été pris en charge du début à la fin de notre séjour. En plus d’avoir pu  réaliser les implants à un prix plus que raisonnable, nous avons eu la chance de pouvoir visiter la magnifique ville d’Antalya. Que demander de plus? J’ai hâte d’y retourner pour la deuxième partie.
+Polarity Scores: {'neg': 0.033, 'neu': 0.936, 'pos': 0.031, 'compound': -0.0258}
+
+Review: Hasta ile ilgi, alaka, güleryüz mükemmel. Tedavimin başlangıç ve bitimine kadar hep yanımda oldular. Profesyonel bir ekip ve çalışma. Tavsiye edebileceğim tek klinik. Herşey için teşekkür ederim
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: Tedavi sürecimde ilgi ve alakalarından çok memnun oldum, yaklaşık 1 senedir kontrollerime gidip geliyorum tedavi sonrası süreçte de çok profesyonel ve başarılılardı. Tüm çalışanlarına teşekkür ediyorum
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: Çok güzel ilgilendiler. Emre Bey'e ilgisinden dolayı ve korkularımı yenmemde yardımcı oldukları için çok teşekkür ederim. Gönül rahatlığıyla gidebilirsiniz.
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
