+++ server/Machine-Learning/Text_Files/Orange_Dental_Turkey_VADER_sentiment.txt
@@ -1,0 +1,28 @@
+Review: My dentist Ziya Osmanli did a great job. He changed two of my dental crowns and made another one for me. He's a professional and just a nice preson. He also speaks English and Russian fluently. Overall, I'm happy with the dentist and the clinic
+Polarity Scores: {'neg': 0.0, 'neu': 0.798, 'pos': 0.202, 'compound': 0.891}
+
+Review: Best experience I've ever had at a dentist I flew from the UK to receive dental implant and could not be happier with my experience the staff are very friendly and careing 100% will recommend to anyone In need of dental care
+Polarity Scores: {'neg': 0.052, 'neu': 0.696, 'pos': 0.252, 'compound': 0.8914}
+
+Review: Hvala divnim ljekarima i medicinskom osoblju na profesionalnosti i ljubaznosti u veoma lijepom ambijentu. Topla preporuka za sve stomatološke usluge.
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: Je remercie tous les docteurs et les staff médical dans c'est clinique,  le service et les rendez-vous,  ainsi les conditions de la visite médicale été a haut niveau, je vous recommande d'aller à cette clinique pour tous les problèmes dentaires.
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: از خدمات که ارائه میکنند بسیار متشکر هستم واقعا دندان پزشکان متخصص در این عرصه کار میکنند و با شما همکاری می‌کنند برای کسانی که Hollywood smile  و امپلانت میخواهند برای شأن این جا را پیشنهاد میکنم
+Polarity Scores: {'neg': 0.0, 'neu': 0.935, 'pos': 0.065, 'compound': 0.3612}
+
+Review: Очень грамотные врачи, делают свою работу на все 100% Ни разу не пожалела, что выбрала эту клинику!
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: Highly recommended
+لم أحصل على تجربة أروع من ذلك شكرا لكل الأطباء
+Polarity Scores: {'neg': 0.0, 'neu': 0.84, 'pos': 0.16, 'compound': 0.2716}
+
+Review: C vraiment top du top ,je vous le recommande vivement
+Polarity Scores: {'neg': 0.0, 'neu': 0.69, 'pos': 0.31, 'compound': 0.3818}
+
+Review: 
+Polarity Scores: {'neg': 0.0, 'neu': 0.0, 'pos': 0.0, 'compound': 0.0}
+
