+++ server/Machine-Learning/Text_Files/Orange_Dental_Turkey_sentiment.txt
@@ -1,0 +1,28 @@
+Review: My dentist Ziya Osmanli did a great job. He changed two of my dental crowns and made another one for me. He's a professional and just a nice preson. He also speaks English and Russian fluently. Overall, I'm happy with the dentist and the clinic
+Sentiment: Sentiment(polarity=0.32857142857142857, subjectivity=0.40714285714285714)
+
+Review: Best experience I've ever had at a dentist I flew from the UK to receive dental implant and could not be happier with my experience the staff are very friendly and careing 100% will recommend to anyone In need of dental care
+Sentiment: Sentiment(polarity=0.74375, subjectivity=0.475)
+
+Review: Hvala divnim ljekarima i medicinskom osoblju na profesionalnosti i ljubaznosti u veoma lijepom ambijentu. Topla preporuka za sve stomatološke usluge.
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
+Review: Je remercie tous les docteurs et les staff médical dans c'est clinique,  le service et les rendez-vous,  ainsi les conditions de la visite médicale été a haut niveau, je vous recommande d'aller à cette clinique pour tous les problèmes dentaires.
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
+Review: از خدمات که ارائه میکنند بسیار متشکر هستم واقعا دندان پزشکان متخصص در این عرصه کار میکنند و با شما همکاری می‌کنند برای کسانی که Hollywood smile  و امپلانت میخواهند برای شأن این جا را پیشنهاد میکنم
+Sentiment: Sentiment(polarity=0.3, subjectivity=0.1)
+
+Review: Очень грамотные врачи, делают свою работу на все 100% Ни разу не пожалела, что выбрала эту клинику!
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
+Review: Highly recommended
+لم أحصل على تجربة أروع من ذلك شكرا لكل الأطباء
+Sentiment: Sentiment(polarity=0.16, subjectivity=0.5399999999999999)
+
+Review: C vraiment top du top ,je vous le recommande vivement
+Sentiment: Sentiment(polarity=0.5, subjectivity=0.5)
+
+Review: 
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
