+++ server/Machine-Learning/Text_Files/Premium_Dental_Turkey_VADER_sentiment.txt
@@ -1,0 +1,64 @@
+Review: Nothing short of AMAZING!    I went to  Premium Dental turkey for my full smile transformation. I was received at the airport with their transportation services free of charge, to and from each appointment. I am forever grateful and thankful for the team at premium dental. They went above and beyond! The entire team and staff were very caring and accommodating, dentists were thorough and professional throughout the entire process explaining everything to me making sure I understood every step. I felt in great hands    definitely recommend PDT if you are looking to get some work done.  I left feeling confident with my new smile
+Polarity Scores: {'neg': 0.028, 'neu': 0.7, 'pos': 0.272, 'compound': 0.983}
+
+Review: Impecable 5* Service from all of the Team at PDT   .. I’m so pleased I chose these guys & gals    ~ they’ve not only gave me an awesome smile.. most importantly, they’ve given me my confidence back.
+
+I feel amazing   
+
+Thank You Premium Dental Turkey
+Polarity Scores: {'neg': 0.0, 'neu': 0.628, 'pos': 0.372, 'compound': 0.9703}
+
+Review: I am delighted with my teeth and the professional and friendly ambience of the clinic. My entire new set of teeth are absolutely stunning and very comfortable. My Dentist in London is so impressed with the work he is referring patients to the clinic. Oceans of love and thanks for looking after me and giving me a smile I always wanted. Simin
+Polarity Scores: {'neg': 0.0, 'neu': 0.647, 'pos': 0.353, 'compound': 0.9817}
+
+Review: I recently had veneers done at Premium Dental and I can confidently say that they did an excellent job. I highly recommend them to anyone looking for high-quality dental work. The staff was amazing - thoughtful, friendly, and attentive. Special thanks to Maddy, the patient coordinator, who took care of all the logistics, including airport pickup/dropoff and transportation to their office every day. She is extremely good at her job and really ensures that your experience there is premium. The drivers were never late, and always friendly, which made the trip so stress-free.
+
+A special thank you to Dr. Ismail who completely transformed my teeth. He not only fixed my smile, but it felt like he gave me a new outlook on life and the confidence I was missing. His attention to detail and expertise is truly impressive. The office itself was the epitome of luxury - beyond clean and modern, it truly represented the premium experience they promised. Overall, my experience at Premium Dental was outstanding and I would refer them to everyone. Thank you for the great work and exceptional service!
+Polarity Scores: {'neg': 0.012, 'neu': 0.686, 'pos': 0.302, 'compound': 0.9961}
+
+Review: I recommend this clinic because after 10 years of waiting I can enjoy a perfect smile.
+I can say that I am a person who is afraid of doctors, but Mady and the entire team of doctors were so patient with me that I ended up feeling very comfortable when I went for treatment.
+So if someone wants perfect teeth, I recommend Premium Dental Turkey
+Polarity Scores: {'neg': 0.0, 'neu': 0.713, 'pos': 0.287, 'compound': 0.9678}
+
+Review: The Clinic received me very well, and the whole Team including the drivers were extremely Professional. My Procedure was done in only 4 Days, and the Team had to Work extra Late just so I can get my veneers in time. I can only recommend this Clinic, They will really take good care of you throughout the whole Procedure. I would highly recommend getting your Veneers or Crowns at this Clinic.
+Polarity Scores: {'neg': 0.0, 'neu': 0.822, 'pos': 0.178, 'compound': 0.9236}
+
+Review: I went to PDT recently for a full treatment and I couldn’t be more happy!
+All the personnel were super friendly and very professional.
+Any doubt or fear was completely gone after the first minute.
+I got 24 zirconium crowns and my teeth never looked this good.
+For anyone that wants a complete makeover and a perfect smile should definitely visit.
+Polarity Scores: {'neg': 0.073, 'neu': 0.637, 'pos': 0.29, 'compound': 0.9565}
+
+Review: I admire their professionalism, their thoroughness in explaining dental procedures and approachability should you have any questions. But what really stands out is their caring demeanour. It’s a quality that is rare and helps to make you feel that you’re in good hands.  Gloria Thomas, writer, children books - London
+Polarity Scores: {'neg': 0.0, 'neu': 0.772, 'pos': 0.228, 'compound': 0.9274}
+
+Review: A+++ for the team and job undertaken! I had 8 upper and 8 lower crowns put on last year september, it is now 1 year past and they are still in same perfect condition! The staff from transfer to actual dental treatment is amazing, everybody makes you feel as part of a family, they look after you like you are VIP       my dentist was Dr. Aras. I had them checked at my local private dentist back in UK and they confirmed that the job is done perfectly! Don’t hesitate to use Premium dental clinic, they are best for such important life changing make over!
+Polarity Scores: {'neg': 0.034, 'neu': 0.762, 'pos': 0.204, 'compound': 0.9723}
+
+Review: My Premium Dentel review:
+I had my treatment back in July, I wanted to wait a couple of months before writing my review to give a full roundup of my experience.
+
+Contact prior to travelling: Excellent.
+By email and WhatsApp, the contact was superb, all questions were answered and explained, I felt total confidence in what the process would be when we arrived.
+
+I decided to book a 10 day TUI holiday so that I was having a trip as well as my treatment.
+Mum and I arrived at the airport and the surgery has arranged for transfers to take us to the hotel, this is a nice touch that they offer.
+The following day we were picked up at the agreed time and taken to the surgery:
+Monday: This is the long day, we were greeted by Maddi, she was extremely friendly, helpful and informative.
+Then the dentists (all 3 of them) examined me, they consulted each other and then Maddi explained the results, the treatments and cost, I was very impressed that they quoted me what treatment would be best for me and not the most expensive.
+Mum was shown to the very nice waiting room and my treatment started, the filling of the teeth, this was quite harrowing, it wasn’t painful due to the injections, it is loud and long, what helps is the hygienist and dentist were very attentive, after hours of filing you wait for you temporary teeth.
+When they come they feel a little odd however you’re just grateful to leave and go back to the hotel at this point.
+I was quite nervous about eating so I stuck to soft foods, but I believe I would have been ok with most things, these teeth were ok but feel a little awkward.
+Thursday, Friday, Saturday: I went back for my fittings, everyday we were met by Maddi who continued to be helpful and supportive.
+The dentist was excellent, he explained everything he was doing as he did it.
+He was a complete perfectionist, which made me fell secure.
+
+Inserting the implants was the most painful day, this is due to the injections and procedure, following this I found it really difficult to get used to my face, I looked completely different, seeing myself in photos freaked me out, however I am now used to my new smile and I LOVE IT, I’ve had great feedback from friends and colleagues.
+
+I had discomfort and sensitivity for about 1 month following, I also had bleeding for a couple of weeks, this stopped when started using Corsodyl toothpaste and mouthwash.
+
+Overall, it was a wonderful service, extremely professional, very clean and worth every penny.
+Polarity Scores: {'neg': 0.051, 'neu': 0.783, 'pos': 0.166, 'compound': 0.9957}
+
