+++ server/Machine-Learning/Text_Files/Queen_Dentaire_Turquie_Easy_Dentaire_Tourisme_dentaire_implant_dentaire_turquie_clinique_Dental_Centre_Turkey_VADER_sentiment.txt
@@ -1,0 +1,18 @@
+Review: Good value and good work. I'm glad I came and didn't have the treatment done in uk
+Polarity Scores: {'neg': 0.0, 'neu': 0.537, 'pos': 0.463, 'compound': 0.8807}
+
+Review: It was a good experience with doctor abdallah kaza he was perfect and very profional
+Polarity Scores: {'neg': 0.0, 'neu': 0.663, 'pos': 0.337, 'compound': 0.765}
+
+Review: Alpi diş merkeziyle tanıştığım için çok mutluyum kaynarcada bayan giyim mağazası işletiyorum gelen müşterilerime şiddetle tavsiye ediyorum hem hijyenik hem güler yüzlü çalışanları tedavimi yapan Dr Ezgi hanımdan Danışman Bahar hanımdan Dr asistanı İrem hanımdan çok memnun kaldım diş bölümünden korkan biriyim fakat bu korkumu Ezgi hanım hem güler yüzü hem kişilik karakter hem tecrübesiyle yenmemi sağladı eşimi kızımı arkadaşlarımı  Tecrübesine bilgisine güvendiğim Ezgi hanıma getirecegim gelen müşterilerimide Ezgi hanıma yönlendireceğim böyle tecrübeli hastaya şefkatle yaklaşan güven veren  biriyle tanıştırdıgı için alpi diş merkezine saygılarımı teşekkürlerimi iletmek istedim ilk defa gidecek olan arkadaşlara tavsiyem gözü kapalı rahatlıkla gidip kendilerini sağlıklarını Dr Ezgi hanıma teslim edebilirler sonuçları mutlu ayrilacaklarina eminim Teşekkürler Dr Ezgi hanım Teşekkürler Bahar hanım Teşekkürler İrem hanım harikasınız
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: Superbe expérience bravo a toute l'équipe  Rassurante . Et respect a tout le personnel depuis la prise en mains. Un grand merci a lequipe dentakay
+Polarity Scores: {'neg': 0.051, 'neu': 0.743, 'pos': 0.206, 'compound': 0.6808}
+
+Review: Çok güler yüzlü bir ekip bütün çalışanlarına ve hocalarına teşekkür ederim…
+Polarity Scores: {'neg': 0.0, 'neu': 1.0, 'pos': 0.0, 'compound': 0.0}
+
+Review: 
+Polarity Scores: {'neg': 0.0, 'neu': 0.0, 'pos': 0.0, 'compound': 0.0}
+
