+++ server/Machine-Learning/Text_Files/Queen_Dentaire_Turquie_Easy_Dentaire_Tourisme_dentaire_implant_dentaire_turquie_clinique_Dental_Centre_Turkey_sentiment.txt
@@ -1,0 +1,18 @@
+Review: Good value and good work. I'm glad I came and didn't have the treatment done in uk
+Sentiment: Sentiment(polarity=0.6333333333333333, subjectivity=0.7333333333333334)
+
+Review: It was a good experience with doctor abdallah kaza he was perfect and very profional
+Sentiment: Sentiment(polarity=0.6333333333333333, subjectivity=0.6333333333333334)
+
+Review: Alpi diş merkeziyle tanıştığım için çok mutluyum kaynarcada bayan giyim mağazası işletiyorum gelen müşterilerime şiddetle tavsiye ediyorum hem hijyenik hem güler yüzlü çalışanları tedavimi yapan Dr Ezgi hanımdan Danışman Bahar hanımdan Dr asistanı İrem hanımdan çok memnun kaldım diş bölümünden korkan biriyim fakat bu korkumu Ezgi hanım hem güler yüzü hem kişilik karakter hem tecrübesiyle yenmemi sağladı eşimi kızımı arkadaşlarımı  Tecrübesine bilgisine güvendiğim Ezgi hanıma getirecegim gelen müşterilerimide Ezgi hanıma yönlendireceğim böyle tecrübeli hastaya şefkatle yaklaşan güven veren  biriyle tanıştırdıgı için alpi diş merkezine saygılarımı teşekkürlerimi iletmek istedim ilk defa gidecek olan arkadaşlara tavsiyem gözü kapalı rahatlıkla gidip kendilerini sağlıklarını Dr Ezgi hanıma teslim edebilirler sonuçları mutlu ayrilacaklarina eminim Teşekkürler Dr Ezgi hanım Teşekkürler Bahar hanım Teşekkürler İrem hanım harikasınız
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
+Review: Superbe expérience bravo a toute l'équipe  Rassurante . Et respect a tout le personnel depuis la prise en mains. Un grand merci a lequipe dentakay
+Sentiment: Sentiment(polarity=0.5, subjectivity=1.0)
+
+Review: Çok güler yüzlü bir ekip bütün çalışanlarına ve hocalarına teşekkür ederim…
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
+Review: 
+Sentiment: Sentiment(polarity=0.0, subjectivity=0.0)
+
