+++ server/Machine-Learning/Text_Files/Özdemir_Dental_Center_Antalya_sentiment.txt
@@ -1,0 +1,49 @@
+Review: Rated 5 out of 5 stars
+Good day, everyone   
+I've just finished my teeth done at this clinic. Now,thanks to them, I have an amazing smile.
+I just want to mention a few things, Miss Meltem was very professional and guided me in perfect English, step by step all the procedures, and explained to me all my questions and my concerns.
+Mrs. Esra is a perfectionist and very professional as well, plus the fact that she is an angel    with all her patience that she had with me.
+And of course, the last but not the least, Mr. Erdem, just amazing and a true master of dentistry.
+Overall, I felt welcomed and treated with respect by everyone at the clinic.
+Thank you all for an amazing job
+Sentiment: Sentiment(polarity=0.2792857142857143, subjectivity=0.41976190476190484)
+
+Review: Just finished my treatment with Ozdemir Dental Centre.  I can honestly say that they are all highly professional and friendly.  The process from start to finish was very smooth. I can recommend them 100%
+Sentiment: Sentiment(polarity=0.39875, subjectivity=0.5375)
+
+Review: Great service
+I visited the clinic in August this year while we’ve been on holiday in Antalya. Beautiful clinic with really friendly people.
+Few days ago they’ve been to Liverpool Dr Erdem and few more colleagues. We met them for consultation. Very well organised team in a nice hotel     
+Even small entertainment place in corner of room for my 2 little boys   
+Hopefully we( me and my wife)can go to Antalya soon for treatment
+Thanks again Dr Erdem and Büşra and Batur bay and others
+Sentiment: Sentiment(polarity=0.24431818181818185, subjectivity=0.4863636363636364)
+
+Review: This place is first class.
+A big thank you to ursin and a very special assistent.
+(Izmet). Keep up the good work as i will recomend you to anyone needing medical treatment.
+Sentiment: Sentiment(polarity=0.2828571428571428, subjectivity=0.35523809523809524)
+
+Review: Booked a laser teeth whitening treatment for my Türkiye holiday. The treatment was quite straight forward and the place was very friendly and professional.
+
+Meltem who works there was very sweet and helpful!
+Sentiment: Sentiment(polarity=0.3390625, subjectivity=0.49875)
+
+Review: A great dental clinic with amazing , friendly and well trained staff. Honest consultation and a very quality job. Went in for a filling and scaling, then i ve been told that i do not need to fill it in as it is good enough. That is a very high level of business ethics and professionalism. The dental job was quick, comfortable , clean and at an amazing price. I totally recommend this clinic to everyone specially , international visitors and expats.
+Sentiment: Sentiment(polarity=0.3702142857142857, subjectivity=0.6287142857142857)
+
+Review: Me and my husband arrived home from Turkey on Saturday after visiting Ozdemir Dental centre. We stayed at Tema 242 for our stay and was more than happy with the free accommodation.
+I had 24 zirconium crowns fitted, before arriving in turkey I was absolutely terrified at the thought of having my teeth shaved down but was put at complete ease from the moment we entered the dental centre. All the staff were absolutely wonderful and so friendly. I can’t thank Batur enough for helping us through the whole process.
+
+I couldn’t be happier with my new smile and highly recommend Ozdemir dental to everyone. A pain free positive experience.
+Sentiment: Sentiment(polarity=0.3095387205387206, subjectivity=0.5819259259259258)
+
+Review: Over the years I have had various dental procedures carried out by Dr Erdem to the highest standard. The clinic is very modern and hygienic with the most up to date equipment. All of the clinic staff are very professional and friendly. My latest course of treatment was with Dr Esra and Dr Erdem. I had deep cleaning, whitening and afterwards Dr Esra replaced 2 very old crowns and in addition I had composite bonding to 8 of my upper teeth which has transformed my smile. I can honestly say it was completely painless, the whole procedure was completed within the time frame and I am so thrilled with the results which are beautiful and very natural. The prices are very reasonable and I highly recommend this clinic.
+Sentiment: Sentiment(polarity=0.2681578947368421, subjectivity=0.4694736842105263)
+
+Review: Amazing service from start to finish, a dentist where they genuinely care for you and your teeth not just what looks best/costs more. All the staff are so friendly and Helpful. Any issues are dealt with promptly the service they provided is outstanding. I recommend this place to everyone it’s really affordable, clean and their patient care is impeccable.
+Sentiment: Sentiment(polarity=0.46145833333333336, subjectivity=0.615625)
+
+Review: Glad to have found this clinic on facebook. Batur was excellent in communicating with me and organising my travel from airport to hotel to clinic. Dr Melisa and her assistant Bushra did a fantastic job on the the zirconia crowns I got there. Thank all for your good work!
+Sentiment: Sentiment(polarity=0.69375, subjectivity=0.875)
+
