+++ server/Machine-Learning/keyword_extraction.py
@@ -1,6 +1,6 @@
 # this file will be used to extract the keywords that was used to cateogirze the sentiment analysis model, this will utilize a seperate machine learning model of it's own
 # this file will be used to extract the keywords that was used to categorize the sentiment analysis model
-from sentiment_analysis import all_companies_sentiments, review_snippets_2d
+from sentiment_analysis1 import all_companies_sentiments, review_snippets_2d
 import os
 import yake
